Connect frontend with backend POST and PUT API

+++ frontend/contact-app-frontend/src/app/components/contactComponents/contact-delete/contact-delete.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
 import { ContactService } from '../../../services/contactService/contact-service';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, FormControl, Validators, FormGroup, ReactiveFormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-contact-delete',
-  imports: [RouterModule, FormsModule],
+  imports: [RouterModule, FormsModule, ReactiveFormsModule, CommonModule],
   templateUrl: './contact-delete.html',
   styleUrl: './contact-delete.css'
 })
@@ -16,20 +17,28 @@ export class ContactDelete {
               private route: ActivatedRoute) { }
 
   contactId: string = "";
-  password: string = "";
+  
+  deleteContactForm = new FormGroup({
+    password: new FormControl('', [Validators.required, Validators.minLength(8), Validators.pattern('^(?=.*[a-z])(?=.*[A-Z])(?=.*\\W)\\S+$')])
+  }); 
 
   ngOnInit(): void{
     this.contactId = this.route.snapshot.paramMap.get('id') ?? "";
   }
 
   onSubmit(): void {
+    if (this.deleteContactForm.invalid) {
+      console.error('Form is invalid');
+      return;
+    }
+
     const deleteRequest = {
-      password: this.password
+      password: this.deleteContactForm.get('password')?.value ?? ""
     }
 
     this.contactService.deleteContact(this.contactId, deleteRequest).subscribe({
       next: (response) => {
-        console.log('Contact deleted successfully:', response);
+        console.log(response);
         this.router.navigate(['/contacts']);
       },
       error: (error) => {
